+++ CHANGELOG.md
@@ -22,6 +22,10 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Nothing
 -->
 
+## UNRELEASED
+### Added
+- Implement FilterableInterface from `spaceonfire/criteria` by collection
+
 ## [1.3.0] - 2020-03-07
 ### Added
 - New methods added:
@@ -33,7 +37,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
     - `CollectionInterface::lastKey()`
     - `CollectionInterface::average()`
     - `CollectionInterface::median()`
-- Method aliases adde:
+- Method aliases added:
     - `BaseCollection::avg()` alias to `BaseCollection::average()`
     - `BaseCollection::join()` alias to `BaseCollection::implode()`